+++ e2e/clients/dashboard-tus/app.js
@@ -10,9 +10,16 @@ import '@uppy/dashboard/dist/style.css'
 const companionUrl = 'http://localhost:3020'
 const uppy = new Uppy()
   .use(Dashboard, { target: '#app', inline: true })
-  .use(Tus, { endpoint: 'https://tusd.tusdemo.net/files' })
+  .use(Tus, { endpoint: 'https://tusd.tusdemo.net/files', onShouldRetry })
   .use(Url, { target: Dashboard, companionUrl })
   .use(Unsplash, { target: Dashboard, companionUrl })
+
+function onShouldRetry (err, retryAttempt, options, next) {
+  if (err?.originalResponse?.getStatus() === 418) {
+    return true
+  }
+  return next(err)
+}
 
 // Keep this here to access uppy in tests
 window.uppy = uppy