+++ packages/@uppy/informer/src/TransitionGroup.js
@@ -2,9 +2,8 @@
 /**
  * @source https://github.com/developit/preact-transition-group
  */
-'use strict'
 
-const { Component, cloneElement, h, toChildArray } = require('preact')
+import { Component, cloneElement, h, toChildArray } from 'preact'
 
 function assign (obj, props) {
   return Object.assign(obj, props)
@@ -286,4 +285,4 @@ TransitionGroup.defaultProps = {
   childFactory: identity,
 }
 
-module.exports = TransitionGroup
+export default TransitionGroup