+++ packages/@uppy/svelte/src/components/DragDrop.svelte
@@ -36,4 +36,4 @@
     uppy.setOptions(options)
   }
 </script>
-<div bind:this={container} />
+<div class="uppy-Container" bind:this={container} />