+++ packages/@uppy/aws-s3/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @uppy/aws-s3
 
+## 2.1.0
+
+Released: 2022-05-14
+Included in: Uppy v2.10.0
+
+- @uppy/aws-s3-multipart,@uppy/aws-s3,@uppy/core,@uppy/react,@uppy/transloadit,@uppy/tus,@uppy/xhr-upload: proposal: Cancel assemblies optional (Mikael Finstad / #3575)
+
 ## 2.0.9
 
 Released: 2022-04-07